v1.2.1 - DB support for users and user creation has been added
This is a major change, so the version has increased by 0.1.0 instead of 0.0.1 like a minor update. This change allows user creation, which shouldve been there from the start????????? Anyways, it was half implemented, and it was with JSON, now its on an actual database, so it'll scale better and perform better too! TODO: User editing, Quiz delete and File Upload